feat: Add MIT License
This commit adds the MIT License to the repository. The license grants permission to use, copy, modify, merge, publish, distribute, sublicense, and/or sell the software, subject to certain conditions. This ensures that the software is provided under an open-source license.

Note: The commit message has been modified to adhere to the established commit message conventions.