feat: Update package name to ecies-bls12381 and add badges
This commit updates the package name in the README.md file to "ecies-bls12381" to reflect the actual name of the package. Additionally, it adds badges for GoDoc, Codecov, Go Report Card, and MIT License to provide important information and improve the visibility of the package.

Note: The commit message has been modified to adhere to the established commit message conventions.